[fix][sw] Update and fix sw for rebase
* Also include changes from quickfix/ila_clock_and_import_ips #157

+++ sw/SoC/examples/custom_hls_conv_hbus/inc/xlnx/driver.h
@@ -6,10 +6,10 @@
 #include "xkrnl_conv_hbus_hw.h"
 
 // Import symbols for peripherals
-extern const volatile uint32_t _peripheral_HLS_CONTROL_start;
+extern const volatile uintptr_t _peripheral_HLS_CONTROL_start;
 
 // Offsets
-#define Xkrnl_BASE             ((unsigned int)(&_peripheral_HLS_CONTROL_start))
+#define Xkrnl_BASE             ((uintptr_t)(&_peripheral_HLS_CONTROL_start))
 #define Xkrnl_Control          (Xkrnl_BASE + XKRNL_CONV_HBUS_CONTROL_ADDR_AP_CTRL)
 #define Xkrnl_GIE              (Xkrnl_BASE + XKRNL_CONV_HBUS_CONTROL_ADDR_GIE)
 #define Xkrnl_IER              (Xkrnl_BASE + XKRNL_CONV_HBUS_CONTROL_ADDR_IER)
